Update patch set 2
Patch Set 2:

(3 comments)

How's this for an interface to these streams? The size is required up front to allow allocating them to a file position, but otherwise this is modeled on MemorySnapshot.
Note the rest of the CL has not been updated to suit.

Patch-set: 2
Reviewer: Sigurður Ásgeirsson <1177549@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>