Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Hi, thanks for the patch. The code changes generally look reasonable, but I'm not sure I generally like letting the target process do more things after crashing. (I know this is what Breakpad used to do though.)

I agree. Putting this on the crashed process is a non-starter.

> I wonder if instead we could have the client (optionally) provide enough data to allow Crashpad to handle the restart?

What’s the restart interface supposed to look like?

Is this something that’s desirable for Chrome? Is it something that we used to have but lost when we switched to Crashpad?

Patch-set: 2
Reviewer: Mark Mentovai <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>