Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > Hi, thanks for the patch. The code changes generally look reasonable, but I'm not sure I generally like letting the target process do more things after crashing. (I know this is what Breakpad used to do though.)
> 
> I agree. Putting this on the crashed process is a non-starter.
> 
> > I wonder if instead we could have the client (optionally) provide enough data to allow Crashpad to handle the restart?
> 
> What’s the restart interface supposed to look like?
> 
> Is this something that’s desirable for Chrome? Is it something that we used to have but lost when we switched to Crashpad?

Yes, Chrome Windows used to put up a dialog on browser crashes that prompted the user to "restart" (I believe the dialog was actually chrome, already restarted). This was implemented by allowing the crashed process to resume after the dump was captured, having it do a CreateProcess() with an environment variable set notifying the new child to show the dialog.

Some references:

https://cs.chromium.org/chromium/src/chrome/browser/chrome_browser_main_win.cc?rcl=fb5af30ea5537cc1f02265dcec0981c1bf10a8d2&l=402 

https://cs.chromium.org/chromium/src/chrome/app/chrome_crash_reporter_client_win.cc?rcl=71125ff12bb0d87b36c81a2783223489bd062a42&l=302

https://cs.chromium.org/chromium/src/components/crash/content/app/breakpad_win.cc?rcl=fb5af30ea5537cc1f02265dcec0981c1bf10a8d2&l=364


I'm a little skeptical of the value of the feature personally, but if we do want to try implementing it I was hoping we could instead push the CreateProcess() arguments into the handler somehow so that it could do the restart after terminating the crashed process. One complexity there is that there's Chrome-specific logic run at crash time (e.g. to avoid restart loops).

Patch-set: 2
Reviewer: Scott Graham <1144260@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>