Update patch set 2
Patch Set 2:

> Yes, Chrome Windows used to put up a dialog on browser crashes that prompted the user to "restart" (I believe the dialog was actually chrome, already restarted). This was implemented by allowing the crashed process to resume after the dump was captured, having it do a CreateProcess() with an environment variable set notifying the new child to show the dialog.
> 
> Some references:
> 
> https://cs.chromium.org/chromium/src/chrome/browser/chrome_browser_main_win.cc?rcl=fb5af30ea5537cc1f02265dcec0981c1bf10a8d2&l=402 
> 
> https://cs.chromium.org/chromium/src/chrome/app/chrome_crash_reporter_client_win.cc?rcl=71125ff12bb0d87b36c81a2783223489bd062a42&l=302
> 
> https://cs.chromium.org/chromium/src/components/crash/content/app/breakpad_win.cc?rcl=fb5af30ea5537cc1f02265dcec0981c1bf10a8d2&l=364
> 
> 
> I'm a little skeptical of the value of the feature personally

On macOS, we get an app-crashed dialog box from the OS. It’s a little clunky, but it is nice to have a button that prompts the user to relaunch. (Plus, that mechanism is also what’s responsible for writing OS crash reports, and when we were still on Breakpad, it was still valuable to have that fallback option.)

> but if we do want to try implementing it I was hoping we could instead push the CreateProcess() arguments into the handler somehow so that it could do the restart after terminating the crashed process. One complexity there is that there's Chrome-specific logic run at crash time (e.g. to avoid restart loops).

Maybe this is something that we could use something like Siggi’s https://chromium-review.googlesource.com/463746/ for once it lands (probably later today).

In fact, this could probably even hook right into that as-is, get a callback while dump generation is in progress, and tell it that you don’t have anything that you want to contribute to the dump. But I think it’d be better to hang this off of a distinct “done” callback. It’d be better for stability (let’s make sure that the dump gets written before trying anything cute) and probably better for the overall system (let’s make sure that the old Chrome is gone before we ask about starting a new one).

Whatever Chrome-specific logic is needed could then just hang out on the Chrome side, in the callback provided by components/crash/content/app/crashpad_win.cc (or delegated from there to chrome/app/chrome_crash_reporter_client_win.cc).

Patch-set: 2
Reviewer: Mark Mentovai <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>