Update patch set 10
Patch Set 10:

> Patch Set 10:
> 
> > Patch Set 10:
> > 
> > (15 comments)
> > 
> > I added tests for reading Self with a large maps file and reading the maps file of a running child process. This hasn't turned up any errors. The phone I've been testing on doesn't have a [vsyscall]/[vectors] entry, but my goobuntu desktop does.
> > 
> > The MapRunningChild test relies on (bad) luck to check for failures, so I also did some tweaking to Initialize to see if I could inject some bad luck. I tried mmapping some memory after reading the vsyscall entry and still didn't end up finding duplicate entries, so this seems to be fixed in recent kernels.
> > 
> > I'm also bypassing the ScopedMmap API once again and using mprotect directly so I can specify protections for individual pages. If these tests stick, do you support adding the ability to change protections on individual pages to ScopedMmap?
> 
> Scratch that about MapRunningChild not turning up errors. It is actually flaky, so it looks like reading the maps file of a still running/actively mapping process can cause entries to be missed. Still investigating.

I'm pretty sure I've pinned down the reason for the missing entries.

http://elixir.free-electrons.com/linux/v4.4/source/fs/proc/task_mmu.c#L166
When the kernel attempts to start/restart a read from the maps file, it tries to pick up where it left off by saving the start address of the last vm_area it wrote out. If it can't find that vm_area (probably because it's been deallocated) then it falls back to counting the number of entries it has already written out and skipping ahead that many entries. Since there are fewer existing entries than we saw previously, this skips over unseen entries too. It seems to me it would be better to find the next vm_area after the last_address, but sadly that doesn't happen. This could also result in repeated entries, but I haven't actually observed that yet. As some extra verification of the cause of these failures, I tested modifying the MapRunningChild test to allocate the tested entries in a low region of memory (with MAP_32BIT), so that the extra allocations the child is working on come after the important ones in the maps file. This made the MapRunningChild test always pass.

Possible resolutions:
1) Do nothing. Provide a warning that you may miss out on data if your target process is running. Expect that these kinds of errors are probably rare in real life; the last entry we read when filling the buffer has to be deallocated in the time between reads.

2) Repeat reading the maps file some number of times to reduce the probability of missing important (i.e. existing before the call to Initialize) entries. As is, the MapRunningChild test failed 20 times in 50 tries. If I put the body of Initialize in a 3-iter loop, I get 2 failures in 50 tries.

3) Use lseek to read overlapping sections of file to reduce the probability of failures. We are most likely to miss entries just after a page boundary. So, read a page of data, scroll back half a page, and start another page-read. I've tested that lseek at least works on /proc/pid/maps on my goobuntu system, but not sure if there will be other gotchas with this strategy.

Patch-set: 10
Reviewer: Joshua Peraza <1172779@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>