Update patch set 6
Patch Set 6:

(9 comments)

Thanks for the patch, Roman! Scott invoked me, so I confirmed what he thought I meant and made a pass over the entire change.

In https://crashpad.chromium.org/bug/23, I’m going to be implementing retry. I expect that the desired behavior is for report upload to not be retried when the collection server rejects it for quota reasons. Is this correct? (Perhaps we should discuss.) Does the collection server unambiguously indicate that it doesn’t want to see a particular report again? How so?

Patch-set: 6