Update patch set 1
Patch Set 1:

Welcome back!

I’m not sure that compiler support is “there” yet. We looked at this a few months ago (for all of Chromium, including Crashpad but also more) in https://codereview.chromium.org/2670873002/ which never wound up landing. dcheng summarized the problems in the description and the discussion that ensued.

To make things work, we needed to move the alignas things around. The alignas/packed combination was also troublesome (with MSVC), and I don’t know if we solved that (but please feel free to double-check that).

alignof is probably more uncontroversial if you’re interested in doing this piecemeal.

Patch-set: 1