linux: Add structures to support requesting crash dumps
These structs live in registration_protocol.h because they will be a
necessary part of the protocol, but they alone do not yet fully
support a registration protocol. Instead, these structs will first be
used by passing a stringified CrashDumpRequest via argv to a new
handler process which will create a dump for a single client and then
exit.

Bug: crashpad:30
Change-Id: I3e1a49cd0906d443914d1a998460e96bdfad88ee