Update patch set 3
Patch Set 3:

When we fork a single-use process to be a ptrace broker, it’s only going to ptrace its parent. So I think that’s right, we shouldn’t need to give it a pid, it’ll always use its ppid, but we will need to give it an address to pick up some information. (And it’ll also need a socket fd back to the main crashpad_handler, but that’ll probably already exist, so we’ll just need to arrange to let the broker know what the fd number is and let it be inherited.)

Patch-set: 3
Reviewer: Mark Mentovai <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>