Update patch set 24
Patch Set 24:

I'm pretty sure there is no requirement that the Ehdr and Phdrs are loaded into memory at all. That is probably why the link_map exist points the debugger directly to the dynamic array, without trying to look the segment up in the program headers.

I've been working on trying to get a custom link script working to test weird executables. It looks like AT_PHDR is set with assumption that the ELF file is mapped at the base address and I want to check that funky program headers can mess that up.

If we do run into problems with modules not having loaded Ehdrs/Phdrs, we could bypass them for shared libraries and just go straight to the dynamic arrays to get symbol information. Figuring out the main executable's dynamic array without headers might be tricky though.

I wouldn't count on finding _r_debug in the loader's dynamic symbol table. It is in the dynamic symbol table in the loader on my goobuntu machine (and declared extern in link.h), but it's not in the dynamic symbol table for the loader on my phone and not declared in the NDK.

One strategy I had for finding the base address for modules/executables is once we have the address of *something* mapped for that module, e.g. the dynamic array, we can look up the name of the mapping for that address and then find the lowest address mapping with the same file name backing it.

Patch-set: 24
Tag: mailMessageId=<CAFRSFxK5f9BMfZydQvee+7bEhS+M9oEsuJoyViWyRi7g4fMx3w@mail.gmail.com>