Update patch set 25
Patch Set 25:

(1 comment)

> I've been working on trying to get a custom link script working to test weird executables. It looks like AT_PHDR is set with assumption that the ELF file is mapped at the base address and I want to check that funky program headers can mess that up.

Yup, it looks to me like AT_PHDR will point to where the first Phdr would be mapped if the first PT_LOAD contained it, but if it doesn’t, AT_PHDR can point to memory outside the ELF image, perhaps even unmapped memory. Ouch. That executable’s probably not getting very far off the ground, so we can basically treat this as impossible for the main executable. That’s good, at least.

> If we do run into problems with modules not having loaded Ehdrs/Phdrs, we could bypass them for shared libraries and just go straight to the dynamic arrays to get symbol information. Figuring out the main executable's dynamic array without headers might be tricky though.

Well, we’re told the load bias in link_map::l_addr at least…

Using Dyn by itself can give us access to the dynamic symbol table, which is good for finding the CrashpadInfo according to the exported symbol you’re planning on leaving as an anchor. But we need to be able to get more out of a module to produce a minidump. We need to get the module’s debug ID, which we’d pull from the NT_GNU_BUILD_ID note if present, so we need the Phdrs to look for PT_NOTE. We also need the module’s load address (and loaded size), because that’s the data that goes into the minidump that tells the server what was loaded where, and therefore which module’s symbol file to consult when unwinding and symbolizing.

> I wouldn't count on finding _r_debug in the loader's dynamic symbol table. It is in the dynamic symbol table in the loader on my goobuntu machine (and declared extern in link.h), but it's not in the dynamic symbol table for the loader on my phone and not declared in the NDK.

OK, that’s fine. We don’t have to count on it being there, but in case the main executable’s Dyn table didn’t have a DT_DEBUG, we can fall back to trying to find _r_debug exported from the loader, and maybe we can even get this exported from Bionic’s loader in the future.

> One strategy I had for finding the base address for modules/executables is once we have the address of *something* mapped for that module, e.g. the dynamic array, we can look up the name of the mapping for that address and then find the lowest address mapping with the same file name backing it.

OK, I’m realizing that this may be the least-bad fallback. Although I’d refine it a bit: rather than just the lowest one with the same name, I’d pick the target map entry where we know the first Phdr or Dyn is, and then look for the highest map entry at or below that one mapped from the same file from file offset 0. This is intended to find the right map entry in case someone mmaps the same file elsewhere in memory. And I’d also check the device and inode numbers for a match, not just the name.

And since we have the link_map from r_debug for most modules, we can check its l_addr (load bias) and l_ld (Dyn) against our computed values once we believe we have a module in hand and we’ve interpreted it. That should at least tell us whether we’ve puzzled things out of the map correctly.

Now it’s true that we’ll still be unable to read modules that really don’t keep their Phdrs around, and that kind of stinks. We can leave this alone for now with a bug filed to track it. In the future, we may need to cope with these cases by trying to open the file on disk. This is bad for a whole host of reasons (no permission, file might have changed on disk, file might not even exist anymore, etc.) but it seems that it may be our only shot, as a fallback to the fallback.

Patch-set: 25
Reviewer: Mark Mentovai <1174099@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>