Add files needed to be able to run tests on swarming.
In order to migrate off of the 32-bit builder and have
proper tryservers set up, we need to make sure we can
run the tests under swarming, rather than running them
locally.

This CL adds the swarming_client dependency to the DEPS file,
adds an .isolate file to describe what we need to run the tests,
and tweaks the run_tests.py script slightly to run correctly
regardless of the current working directory.

Bug: 697636, 743139
Change-Id: Iab3136fc9a323caf368b54b0b8c5192f5cf29477